requirements.txt
bert-serving-client==1.10.0
bert-serving-server==1.10.0
pandas==0.24.2
Flask==1.1.1
scikit-learn==0.22.1
spacy==2.2.4
requests==2.23.0
numpy==1.18.1
matplotlib==3.2.0
seaborn==0.10.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
spacy-stanfordnlp==0.1.3
sentence-transformers==0.2.6.1
six
nltk
imblearn
tensorflow-hub==0.7.0
gensim
wikipedia
transformers==2.8.0
xlrd
unidecode
networkx
segtok==1.5.11
yake    
#sentence_transformers==1.2.0